+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/extensions/lmbd.py
@@ -23,8 +23,12 @@ from opentelemetry.instrumentation.botocore.extensions.types import (
     _AwsSdkCallContext,
     _AwsSdkExtension,
     _BotocoreInstrumentorContext,
+    _BotoResultT,
 )
 from opentelemetry.propagate import inject
+from opentelemetry.semconv._incubating.attributes.aws_attributes import (
+    AWS_LAMBDA_RESOURCE_MAPPING_ID,
+)
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.trace.span import Span
 
@@ -71,9 +75,13 @@ class _OpInvoke(_LambdaOperation):
     @classmethod
     def _parse_function_name(cls, call_context: _AwsSdkCallContext):
         function_name_or_arn = call_context.params.get("FunctionName")
+        if function_name_or_arn is None:
+            return None
         matches = cls.ARN_LAMBDA_PATTERN.match(function_name_or_arn)
-        function_name = matches.group(1)
-        return function_name_or_arn if function_name is None else function_name
+        if matches:
+            function_name = matches.group(1)
+            return function_name if function_name else function_name_or_arn
+        return function_name_or_arn
 
     @classmethod
     def before_service_call(cls, call_context: _AwsSdkCallContext, span: Span):
@@ -115,6 +123,10 @@ class _LambdaExtension(_AwsSdkExtension):
         self._op = _OPERATION_MAPPING.get(call_context.operation)
 
     def extract_attributes(self, attributes: _AttributeMapT):
+        resource_mapping_id = self._call_context.params.get("UUID")
+        if resource_mapping_id:
+            attributes[AWS_LAMBDA_RESOURCE_MAPPING_ID] = resource_mapping_id
+
         if self._op is None:
             return
 
@@ -127,3 +139,15 @@ class _LambdaExtension(_AwsSdkExtension):
             return
 
         self._op.before_service_call(self._call_context, span)
+
+    def on_success(
+        self,
+        span: Span,
+        result: _BotoResultT,
+        instrumentor_context: _BotocoreInstrumentorContext,
+    ):
+        resource_mapping_id = result.get("UUID")
+        if resource_mapping_id:
+            span.set_attribute(
+                AWS_LAMBDA_RESOURCE_MAPPING_ID, resource_mapping_id
+            )